implemented changes
