+++ src/app/static/src/app/components/adr/SelectDataset.vue
@@ -96,6 +96,7 @@
     import TreeSelect from "@riophae/vue-treeselect";
     import {
         mapActionByName,
+        mapGetterByName,
         mapMutationByName,
         mapStateProp,
     } from "../../utils";
@@ -142,12 +143,12 @@
         updateDatasetRelease: (release: Release) => void;
         updateValid: (valid: boolean) => void;
         preSelectDataset: () => void;
-        checkResourceAvailable: (resourceType: string) => boolean;
+        // checkResourceAvailable: (resourceType: string) => boolean;
     }
 
     interface Computed {
         datasets: any[];
-        selectedDatasetFromDatasets: any | null;
+        // selectedDatasetFromDatasets: any | null;
         selectedRelease: Release | null;
         releaseName: string | null;
         releaseURL: string;
@@ -164,7 +165,8 @@
         disableImport: boolean;
         hasPjnzFile: boolean;
         hasPopulationFile: boolean;
-        availableResources: { [k in keyof DatasetResourceSet]?: DatasetResource | null }
+        selectedDatasetAvailableResources: { [k in keyof DatasetResourceSet]?: DatasetResource | null }
+        // availableResources: { [k in keyof DatasetResourceSet]?: DatasetResource | null }
     }
 
     interface Data {
@@ -242,6 +244,7 @@
                 namespace,
                 (state: ADRState) => state.adrError
             ),
+            selectedDatasetAvailableResources: mapGetterByName("baseline", "selectedDatasetAvailableResources"),
             releaseName() {
                 return this.selectedRelease?.name || null;
             },
@@ -301,24 +304,24 @@
             disableImport() {
                 return !this.newDatasetId || !this.valid
             },
-            selectedDatasetFromDatasets(){
-                return this.datasets.find(dataset => dataset.id === this.newDatasetId) || null
-            },
-            availableResources(){
-                const resourceTypes = {
-                    pjnz: "inputs-unaids-spectrum-file",
-                    pop: "inputs-unaids-population",
-                    shape: "inputs-unaids-geographic",
-                    survey: "inputs-unaids-survey",
-                    program: "inputs-unaids-art",
-                    anc: "inputs-unaids-anc"
-                }
-                const resources: { [k in keyof DatasetResourceSet]?: DatasetResource | null } = {}
-                Object.entries(resourceTypes).forEach(([key, value]) => {
-                    resources[key as keyof typeof resources] = this.checkResourceAvailable(value) ? this.selectedDataset!.resources[key as keyof typeof resources] : null
-                })
-                return resources
-            }
+            // selectedDatasetFromDatasets(){
+            //     return this.datasets.find(dataset => dataset.id === this.newDatasetId) || null
+            // },
+            // availableResources(){
+            //     const resourceTypes = {
+            //         pjnz: "inputs-unaids-spectrum-file",
+            //         pop: "inputs-unaids-population",
+            //         shape: "inputs-unaids-geographic",
+            //         survey: "inputs-unaids-survey",
+            //         program: "inputs-unaids-art",
+            //         anc: "inputs-unaids-anc"
+            //     }
+            //     const resources: { [k in keyof DatasetResourceSet]?: DatasetResource | null } = {}
+            //     Object.entries(resourceTypes).forEach(([key, value]) => {
+            //         resources[key as keyof typeof resources] = this.checkResourceAvailable(value) ? this.selectedDataset!.resources[key as keyof typeof resources] : null
+            //     })
+            //     return resources
+            // }
         },
         methods: {
             getDatasets: mapActionByName("adr", "getDatasets"),
@@ -338,9 +341,9 @@
             importSurvey: mapActionByName("surveyAndProgram", "importSurvey"),
             importProgram: mapActionByName("surveyAndProgram", "importProgram"),
             importANC: mapActionByName("surveyAndProgram", "importANC"),
-            checkResourceAvailable(resourceType){
-                return this.selectedDatasetFromDatasets?.resources.some((resource: any) => resource.resource_type && resource.resource_type === resourceType)
-            },
+            // checkResourceAvailable(resourceType){
+            //     return this.selectedDatasetFromDatasets?.resources.some((resource: any) => resource.resource_type && resource.resource_type === resourceType)
+            // },
             async importDataset() {
                 this.stopPolling();
 
@@ -358,7 +361,7 @@
                     survey,
                     program,
                     anc,
-                } = this.availableResources;
+                } = this.selectedDatasetAvailableResources;
 
                 await Promise.all([
                     pjnz && this.importPJNZ(pjnz.url),
